chore: Update Dockerfile and docker-compose.yaml for ts-meo-discord
This commit updates the Dockerfile and docker-compose.yaml files to configure the project for ts-meo-discord. The WORKDIR in the Dockerfile has been changed to /ts-meo-discord, and the docker-compose.yaml file has been added to define the ts-meo-discord service. These changes ensure that the project is properly set up for running ts-meo-discord in a Docker environment.